send working, fixed loopback issue and added socketcan interface parameter

+++ src/can/can_message.rs
@@ -1,11 +1,11 @@
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub struct CanMessage {
-    pub can_id: u32,
+    pub can_id: u16,
     pub body: Vec<u8>,
 }
 
 impl CanMessage {
-    pub fn new(can_id: u32, body: &[u8]) -> Self {
+    pub fn new(can_id: u16, body: &[u8]) -> Self {
         CanMessage {
             can_id,
             body: Vec::from(body),
